refactor: organize pages into feature directories
