+++ MJ_FB_Frontend/src/App.tsx
@@ -13,18 +13,18 @@ import VolunteerLogin from './pages/auth/VolunteerLogin';
 import VolunteerDashboard from './pages/volunteer/VolunteerDashboard';
 import VolunteerManagement from './pages/volunteer-management/VolunteerManagement';
 import Dashboard from './pages/Dashboard';
-import UserDashboard from './pages/UserDashboard';
+import UserDashboard from './pages/user/UserDashboard';
 import VolunteerBookingHistory from './pages/volunteer/VolunteerBookingHistory';
 import VolunteerSchedule from './pages/volunteer/VolunteerSchedule';
-import WarehouseDashboard from './pages/WarehouseDashboard';
-import DonationLog from './pages/DonationLog';
-import TrackPigpound from './pages/TrackPigpound';
-import TrackOutgoingDonations from './pages/TrackOutgoingDonations';
-import TrackSurplus from './pages/TrackSurplus';
-import Aggregations from './pages/Aggregations';
-import DonorProfile from './pages/DonorProfile';
-import AdminStaffList from './pages/AdminStaffList';
-import AdminStaffForm from './pages/AdminStaffForm';
+import WarehouseDashboard from './pages/warehouse-management/WarehouseDashboard';
+import DonationLog from './pages/warehouse-management/DonationLog';
+import TrackPigpound from './pages/warehouse-management/TrackPigpound';
+import TrackOutgoingDonations from './pages/warehouse-management/TrackOutgoingDonations';
+import TrackSurplus from './pages/warehouse-management/TrackSurplus';
+import Aggregations from './pages/warehouse-management/Aggregations';
+import DonorProfile from './pages/warehouse-management/DonorProfile';
+import AdminStaffList from './pages/admin/AdminStaffList';
+import AdminStaffForm from './pages/admin/AdminStaffForm';
 import Events from './pages/Events';
 import PantryVisits from './pages/staff/PantryVisits';
 import Navbar, { type NavGroup, type NavLink } from './components/Navbar';
