+++ MJ_FB_Frontend/src/pages/user/UserDashboard.tsx
@@ -18,10 +18,10 @@ import {
   DialogActions,
 } from '@mui/material';
 import { EventAvailable, Announcement, History } from '@mui/icons-material';
-import FeedbackSnackbar from '../components/FeedbackSnackbar';
-import { getBookingHistory, getSlots, getHolidays, cancelBooking } from '../api/bookings';
-import type { Slot, Holiday } from '../types';
-import { formatTime } from '../utils/time';
+import FeedbackSnackbar from '../../components/FeedbackSnackbar';
+import { getBookingHistory, getSlots, getHolidays, cancelBooking } from '../../api/bookings';
+import type { Slot, Holiday } from '../../types';
+import { formatTime } from '../../utils/time';
 
 interface Booking {
   id: number;
