+++ MJ_FB_Frontend/src/pages/warehouse-management/DonationLog.tsx
@@ -20,12 +20,12 @@ import {
   IconButton,
 } from '@mui/material';
 import { Edit, Delete } from '@mui/icons-material';
-import Page from '../components/Page';
-import FeedbackSnackbar from '../components/FeedbackSnackbar';
-import { getDonors, createDonor } from '../api/donors';
-import { getDonations, createDonation, updateDonation, deleteDonation } from '../api/donations';
-import type { Donor } from '../api/donors';
-import type { Donation } from '../api/donations';
+import Page from '../../components/Page';
+import FeedbackSnackbar from '../../components/FeedbackSnackbar';
+import { getDonors, createDonor } from '../../api/donors';
+import { getDonations, createDonation, updateDonation, deleteDonation } from '../../api/donations';
+import type { Donor } from '../../api/donors';
+import type { Donation } from '../../api/donations';
 
 function startOfWeek(date: Date) {
   const d = new Date(date);
